requirements_production.txt
# Core Flask API dependencies
Flask==3.0.3
flask-jwt-extended==4.6.0
flask-cors==4.0.0
Werkzeug==3.0.6
gunicorn==21.2.0

# Spotify API
spotipy==2.25.1
requests==2.32.3

# Data processing
pandas==2.2.3
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.0

# Environment and config
python-dotenv==1.1.0

# AI features (optional - remove if not using)
google-generativeai==0.8.3

# Essential utilities
Jinja2==3.1.6
markupsafe==2.1.3
itsdangerous==2.2.0
click==8.2.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0
python-dateutil==2.9.0.post0
typing_extensions==4.13.2
tzdata==2025.2
six==1.17.0

# Note: sqlite3 is built into Python, no need to install
# Removed dash, plotly, and dash-bootstrap-components as they're not needed for API-only backend